requirements.txt
flask
flask-sqlalchemy
flask-cors
pandas
numpy
gunicorn
psycopg2-binary
bcrypt
scikit-learn